0.7.0 update

+++ index.py
@@ -1,3 +1,5 @@
+import modules.single_instance
+
 from datetime import datetime
 import sys
 import time
@@ -7,50 +9,44 @@ import ctypes
 
 from PySide6.QtWidgets import QApplication, QSystemTrayIcon, QMenu, QMessageBox
 from PySide6.QtGui import QIcon
+from PySide6.QtCore import Signal
 
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
 
 from concurrent.futures import ThreadPoolExecutor, wait
 
-from modules.config import get_config
+from modules.config import Config
 from modules.db import MarketDb
+from modules.messagebox import MessageBoxHandler
 from modules.scan import scan
-from modules.sound import playCheck, playError, playSuccess
+from modules.sound import VolumeController, playCheck, playError, playSuccess
 
 from ui.config.config import LostArkMarketWatcherConfig
 from ui.log.log import LostArkMarketWatcherLog
-
-version = '0.6.3'
-debug = False
 
 
 class LostArkMarketWatcher(QApplication):
     market_db = None
     observer = None
     config_form = None
-    play_audio = True
-    save_log = False
-    delete_screenshots = True
-    screenshots_directory = None
     screenshot_executor = None
+    message_box = Signal(dict)
+    message_box_handler: MessageBoxHandler
 
     def __init__(self, *args, **kwargs):
         QApplication.__init__(self, *args, **kwargs)
         self.setQuitOnLastWindowClosed(False)
         self.build_menu()
-        self.market_db = MarketDb(version)
-        self.log_view = LostArkMarketWatcherLog(version, self.market_db.region)
-        self.config_form = LostArkMarketWatcherConfig(
-            version, self.market_db.region)
+        self.market_db = MarketDb()
+        self.message_box_handler = MessageBoxHandler(self.message_box)
+        self.log_view = LostArkMarketWatcherLog()
+        self.config_form = LostArkMarketWatcherConfig()
         self.config_form.config_updated.connect(self.spawn_observer)
         self.market_db.log.connect(self.write_log)
         self.market_db.error.connect(self.write_error)
         self.market_db.new_version.connect(self.new_version)
         self.spawn_observer()
-
-    def get_config(self):
-        self.play_audio, self.delete_screenshots, self.screenshots_directory, self.save_log = get_config()
 
     def build_menu(self):
         icon = QIcon(os.path.abspath(os.path.join(os.path.dirname(__file__),
@@ -73,24 +69,27 @@ class LostArkMarketWatcher(QApplication):
         self.log_view.show()
 
     def open_config(self):
-        self.get_config()
-        self.config_form.show_ui(self.play_audio, self.delete_screenshots,
-                                 self.screenshots_directory, self.save_log)
+        self.config_form.show_ui()
 
     def close_action(self):
         self.tray.hide()
         os._exit(1)
 
     def spawn_observer(self):
-        self.get_config()
-
-        if self.screenshots_directory == None:
-            if self.play_audio == True:
-                playError()
-            self.write_log('Screenshots directory not found')
-            self.open_config()
-            return
-
+        screenshots_directory = None
+        if Config().screenshots_directory:
+            screenshots_directory = Config().screenshots_directory
+        else:
+            if Config().game_directory:
+                screenshots_directory = os.path.abspath(os.path.join(
+                    Config().game_directory, 'EFGame', 'Screenshots'))
+            else:
+                self.write_log('Screenshots directory not found')
+                if Config().play_audio == True:
+                    playError()
+                self.open_config()
+                return
+        print(screenshots_directory)
         event_handler = FileSystemEventHandler()
         event_handler.on_created = self.on_created
 
@@ -100,64 +99,70 @@ class LostArkMarketWatcher(QApplication):
 
         self.observer = Observer()
         self.observer.schedule(
-            event_handler, self.screenshots_directory, recursive=False)
+            event_handler, screenshots_directory, recursive=False)
 
         self.write_log('Watcher Started')
         self.observer.start()
-        self.screenshot_executor = ThreadPoolExecutor(max_workers=1)
+        self.screenshot_executor = ThreadPoolExecutor(
+            max_workers=Config().screenshot_threads)
 
-        if self.play_audio == True:
+        if Config().play_audio == True:
             playSuccess()
 
     def on_created(self, event):
-        self.screenshot_executor.submit(
-            self.process_screenshot, event.src_path)
+        # Region Check
+        Config().get_game_region()
+        if Config().game_region == Config().region:
+            self.screenshot_executor.submit(
+                self.process_screenshot, event.src_path)
+        else:
+            self.message_box.emit({"type": "REGION"})
 
     def process_screenshot(self, file):
         try:
             self.write_log('New screenshoot found')
             time.sleep(2)
-            if self.play_audio == True:
+            if Config().play_audio == True:
                 playCheck()
 
             self.write_log('Scanning: Start')
-            res = scan(file, debug, self.write_log, self.write_error)
+            res = scan(file, self.write_log, self.write_error)
             self.write_log('Scanning: Finish')
             entry_futures = []
-            entries_executor = ThreadPoolExecutor(max_workers=2)
+            entries_executor = ThreadPoolExecutor(
+                max_workers=Config().upload_threads)
             for item in res:
                 entry_futures.append(entries_executor.submit(
-                    self.market_db.add_entry, item, self.play_audio))
+                    self.market_db.add_entry, item))
             wait(entry_futures)
             self.write_log("Finished")
-            if self.play_audio == True:
+            if Config().play_audio == True:
                 playSuccess()
             time.sleep(1)
-            os.remove(file)
+            if Config().delete_screenshots == True:
+                os.remove(file)
         except:
-            playError()
+            if Config().play_audio == True:
+                playError()
             self.write_error(traceback.format_exc())
 
     def write_log(self, txt):
-        self.log_view.log(txt, False, self.save_log)
+        self.log_view.log(txt, False)
 
     def write_error(self, txt):
-        self.log_view.log(txt, True, self.save_log)
-    
+        self.log_view.log(txt, True)
+
     def new_version(self, new_version):
-        msgBox = QMessageBox()
-        msgBox.setWindowTitle("Lost Ark Market Online")
-        msgBox.setText(f"New version v{new_version} available.")
-        msgBox.setInformativeText("Please close the app and run the launcher to download the new version.")
-        msgBox.setIcon(QMessageBox.Information)
-        msgBox.exec()
+        self.message_box.emit({"type": "REGION", "new_version": new_version})
 
 
 if __name__ == "__main__":
-    myappid = f'lostarkmarketonline.watcher.app.{version}'
+    myappid = f'lostarkmarketonline.watcher.app.{Config().version}'
     ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(myappid)
     app = LostArkMarketWatcher([])
     icon = QIcon(os.path.abspath(os.path.join(os.path.dirname(__file__),
-                                                "assets/icons/favicon.png")))
+                                              "assets/icons/favicon.png")))
     app.setWindowIcon(icon)
+    time.sleep(2)
+    VolumeController()
     sys.exit(app.exec())
